pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "onto144-tools"
version = "0.1.0"
description = "Validation and compliance tools for the onto144 ontological registry"
readme = "README.md"
license = {text = "GPL-3.0-only"}
authors = [{name = "onto144 Community", email = "maintainers@onto144.org"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.18.0"
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/onto144-public"
Repository = "https://github.com/onto144-public/onto144-public"
Documentation = "https://github.com/onto144-public/onto144-public/tree/main/docs"

[project.scripts]
validate-profile = "tools.validate_profile:main"
validate-lexicon = "tools.validate_lexicon:main"
check-profile-lexicon-compliance = "tools.check_profile_lexicon_compliance:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tools*"]

[tool.setuptools.package-data]
tools = ["py.typed"]

[tool.black]
line-length = 99
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 99

# Optional: for development environments
[project.optional-dependencies]
dev = [
    "black>=22.0",
    "isort>=5.10",
    "flake8>=5.0",
    "pytest>=7.0"
]